# docs/web-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ajib API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/downloads": {
      "get": {
        "summary": "Downloads",
        "operationId": "downloads_api_v1_downloads_get",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "fa",
                "en",
                "tk",
                "ru"
              ],
              "type": "string",
              "default": "en",
              "title": "Language"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DownloadResponse"
                  },
                  "title": "Response Downloads Api V1 Downloads Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/challenge": {
      "post": {
        "summary": "Challenge",
        "operationId": "challenge_api_v1_auth_challenge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/consume": {
      "post": {
        "summary": "Consume",
        "operationId": "consume_api_v1_auth_consume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/telegram": {
      "post": {
        "summary": "Telegram",
        "operationId": "telegram_api_v1_auth_telegram_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MiniAppInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/language": {
      "put": {
        "summary": "Language",
        "operationId": "language_api_v1_me_language_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguageInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChallengeInput": {
        "properties": {
          "storefront": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "pattern": "^[a-z0-9-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storefront"
          }
        },
        "type": "object",
        "title": "ChallengeInput"
      },
...
      "ConsumeInput": {
        "properties": {
          "challenge": {
            "type": "string",
            "maxLength": 64,
            "minLength": 20,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Challenge"
          }
        },
        "type": "object",
        "required": [
          "challenge"
        ],
        "title": "ConsumeInput"
      },
      "DownloadResponse": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "details": {
            "type": "string",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "id",
          "label",
          "url",
          "details"
        ],
        "title": "DownloadResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LanguageInput": {
        "properties": {
          "language": {
            "type": "string",
            "enum": [
              "fa",
              "en",
              "tk",
              "ru"
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "language"
        ],
        "title": "LanguageInput"
      },
      "MiniAppInput": {
        "properties": {
          "storefront": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "pattern": "^[a-z0-9-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storefront"
          },
          "init_data": {
            "type": "string",
            "maxLength": 16384,
            "minLength": 1,
            "title": "Init Data"
          }
        },
        "type": "object",
        "required": [
          "init_data"
        ],
        "title": "MiniAppInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}